pr14 card to delete by user only completed

+++ controllers/cards.js
@@ -1,7 +1,13 @@
 const Card = require('../models/Card');
 // const res = require('express/lib/response');
+const { isAuthorized } = require('../middlewares/auth');
 
 const getCards = async (req, res) => {
+  const matched = await isAuthorized(req.headers.authorization);
+  if (!matched) {
+    res.status(401).send({ message: 'Нет доступа' });
+    return;
+  }
   try {
     const cards = await Card.find({});
     res.status(200).send(cards);
@@ -14,15 +20,24 @@ const getCards = async (req, res) => {
 };
 
 const deleteCard = async (req, res) => {
+  const matched = await isAuthorized(req.headers.authorization);
+  if (!matched) {
+    res.status(401).send({ message: 'Нет доступа' });
+    return;
+  }
   try {
-    // const deletedCard = await Card.deleteOne({ _id: req.params.cardId });
-    const deletedCard = await Card.findByIdAndRemove(req.params.cardId);
-    if (!deletedCard) {
+    const cardToDelete = await Card.findById(req.params.cardId);
+    if (!cardToDelete) {
       res.status(404).send({
         message: 'Карточки с таким id не найдено',
       });
       return;
     }
+    if (req.user._id !== cardToDelete.owner.toString()) {
+      res.status(403).send({ message: 'Нельзя удалять чужие карточки' });
+      return;
+    }
+    const deletedCard = await Card.findByIdAndRemove(req.params.cardId);
     res.status(200).send(deletedCard);
   } catch (err) {
     if (err.kind === 'ObjectId') {
@@ -37,6 +52,11 @@ const deleteCard = async (req, res) => {
 };
 
 const createCard = async (req, res) => {
+  const matched = await isAuthorized(req.headers.authorization);
+  if (!matched) {
+    res.status(401).send({ message: 'Нет доступа' });
+    return;
+  }
   try {
     const newCard = new Card({
       name: req.body.name,
@@ -54,6 +74,11 @@ const createCard = async (req, res) => {
 };
 
 const likeCard = async (req, res) => {
+  const matched = await isAuthorized(req.headers.authorization);
+  if (!matched) {
+    res.status(401).send({ message: 'Нет доступа' });
+    return;
+  }
   try {
     const likedCard = await Card.findByIdAndUpdate(
       req.params.cardId,
@@ -80,6 +105,11 @@ const likeCard = async (req, res) => {
 };
 
 const dislikeCard = async (req, res) => {
+  const matched = await isAuthorized(req.headers.authorization);
+  if (!matched) {
+    res.status(401).send({ message: 'Нет доступа' });
+    return;
+  }
   try {
     const dislikedCard = await Card.findByIdAndUpdate(
       req.params.cardId,
@@ -112,3 +142,4 @@ module.exports = {
   likeCard,
   dislikeCard,
 };
+// const deletedCard = await Card.deleteOne({ _id: req.params.cardId });
